Re-implement MCP async support on the reworked AsyncToolset API: MCPToolset extends AsyncToolset, non-blocking MCP tools carry ToolFlag.CANCELLABLE and forward progress through ctx.update(), and get_running_tasks/cancel_task are auto-exposed by the activity when a cancellable tool is present rather than added by the toolset.

+++ livekit-plugins/livekit-plugins-google/livekit/plugins/google/tts.py
@@ -77,7 +77,7 @@ class TTS(tts.TTS):
         model_name: NotGivenOr[GeminiTTSModels | str] = NOT_GIVEN,
         prompt: NotGivenOr[str] = NOT_GIVEN,
         sample_rate: int = 24000,
-        pitch: int = 0,
+        pitch: float = 0,
         effects_profile_id: str = "",
         speaking_rate: float = 1.0,
         volume_gain_db: float = 0.0,
